refactor: move sanitizeFilenameSegment to shared filename-utils module
Agent-Logs-Url: https://github.com/KenSuenobu/objectified-commercial/sessions/3ee7cff3-be95-4ffc-84ae-2deaed2687d6
